refactor: userErrorCode 패키지 수정 (#3)
